# WIP: speed up `BatchWrite` by sorting the batches in descending order